test(DES-3210): test method setStatus of class Cell
